Move tf_close_coeff to xferfcn
